# Prometheus: Reduce allocations parsing exemplars