--- a/pkg/tsdb/prometheus/querydata/framing_bench_test.go
+++ b/pkg/tsdb/prometheus/querydata/framing_bench_test.go
@@ -37,6 +37,7 @@ func BenchmarkExemplarJson(b *testing.B) {
 	tCtx, err := setup(true)
 	require.NoError(b, err)
 	b.ResetTimer()
+	b.ReportAllocs()
 	for n := 0; n < b.N; n++ {
 		res := http.Response{
 			StatusCode: 200,